+++ src/ofxGgmlAudio.h
@@ -4,5 +4,6 @@
 #include "ofxGgmlAudio/ofxGgmlAudioTypes.h"
 #include "ofxGgmlAudio/ofxGgmlAudioFeatures.h"
 #include "ofxGgmlAudio/ofxGgmlAudioStreamChunker.h"
+#include "ofxGgmlAudio/ofxGgmlAudioRollingTranscript.h"
 #include "ofxGgmlAudio/ofxGgmlAudioUtils.h"
 #include "ofxGgmlAudio/ofxGgmlAudioWhisperBackend.h"